requirements.txt
flask>=2.0.0
openai>=1.0.0
httpx>=0.24.0
requests>=2.30.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0 